requirements.txt
joblib==1.3.2
numpy==1.26.1
mysql-connector-python==8.1.0
pandas==1.5.3